Move ElementField component to src/components/generic/ folder
